requirements.txt
streamlit==1.32.0
torch==2.1.0
torchvision==0.16.0
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1